requirements.txt
# 数据处理
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# 可视化
matplotlib>=3.4.0
seaborn>=0.11.0

# 优化
scikit-learn>=0.24.0

# 数据存储
h5py>=3.0.0

# 其他工具
tqdm>=4.62.0

# dflow 工作流
pydflow>=1.6.0

# HTTP 服务 (Agent 应用必需)
requests>=2.26.0

# Web 界面
gradio>=4.0.0

